+++ src/metadata/metadata_raw_atomic.c
@@ -94,7 +94,7 @@ static int _raw_atomic_io_discard_do(struct ocf_cache *cache, void *context,
 }
 
 void raw_atomic_flush_mark(struct ocf_cache *cache, struct ocf_request *req,
-		uint32_t map_idx, int to_state, uint8_t start, uint8_t stop)
+		uint32_t map_idx, int to_state, uint32_t start, uint32_t stop)
 {
 	if (to_state == INVALID) {
 		req->map[map_idx].flush = true;
