+++ src/metadata/metadata_raw_dynamic.h
@@ -83,7 +83,7 @@ void raw_dynamic_flush_all(ocf_cache_t cache, struct ocf_metadata_raw *raw,
  * RAW DYNAMIC - Mark specified entry to be flushed
  */
 void raw_dynamic_flush_mark(ocf_cache_t cache, struct ocf_request *req,
-		uint32_t map_idx, int to_state, uint8_t start, uint8_t stop);
+		uint32_t map_idx, int to_state, uint32_t start, uint32_t stop);
 
 /*
  * DYNAMIC Implementation - Do Flush Asynchronously
