+++ src/metadata/metadata_structs.h
@@ -219,8 +219,8 @@ struct ocf_metadata_iface {
 	 * @param[in] line - cache line which to be flushed
 	 */
 	void (*flush_mark)(struct ocf_cache *cache, struct ocf_request *req,
-			uint32_t map_idx, int to_state, uint8_t start,
-			uint8_t stop);
+			uint32_t map_idx, int to_state, uint32_t start,
+			uint32_t stop);
 
 	/**
 	 * @brief Flush marked cache lines asynchronously
